% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafloodr-package.R
\docType{package}
\name{cafloodr-package}
\alias{cafloodr-package}
\title{cafloodr: Tools to Preprocess DEMs and Run the CADDIES/CAFlood Flood Model in R}
\description{
Provides tools for preprocessing Digital Elevation Models (DEMs) and running the 2D cellular automata flood model CADDIES/CAFlood (\url{https://doi.org/10.1016/j.proeng.2016.07.409}) directly from R. Functions include DEM conditioning, rainfall and inflow setup, control point generation, and automated model execution. Designed for flood modeling and hydrodynamic simulation at the catchment scale.
}
\author{
\strong{Maintainer}: Marcos Roberto Benso \email{marcosbenso@gmail.com}

}
\keyword{internal}
