% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{cafloodr_A01}
\alias{cafloodr_A01}
\title{Run CAFlood Hydrological Simulation Version 2}
\usage{
cafloodr_A01(
  param,
  caflood_path,
  Input,
  Output,
  event_name,
  dem_name,
  path2,
  outlet_path = "C:/Projetos/01_DTI-A/02_Input/01_SP/01_Hidro_hora/estacoes.shp",
  rain_path,
  stream_network,
  snap_dist
)
}
\arguments{
\item{param}{List. Model parameters including: `alpha_fraction`, `roughness_global`, `slope_tolerance`, `boundary_ele`, `Raster_WD_Tolerance`, `Upstream_Reduction`.}

\item{caflood_path}{Character. Path to the CAFlood executable.}

\item{Input}{Character. Directory containing model input files.}

\item{Output}{Character. Directory to store model output.}

\item{event_name}{Character. Name or ID of the event; used for file naming.}

\item{dem_name}{Character. Base name of the DEM file (without extension).}

\item{path2}{Character. Subdirectory inside `Input` where the DEM is located.}

\item{outlet_path}{Character. Path to outlet shapefile (default: predefined path).}

\item{rain_path}{Character. Path to rainfall CSV file.}

\item{stream_network}{Character. Path to stream network shapefile (without `.shp` extension).}

\item{snap_dist}{Numeric. Max distance to snap outlets to stream network (meters).}
}
\value{
Logical. `TRUE` if the simulation was triggered successfully, otherwise `FALSE`.
}
\description{
Executes the CAFlood model with specified parameters and input datasets for hydrodynamic flood modeling.
}
