% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_treatment.R
\name{ForwardMole1D}
\alias{ForwardMole1D}
\title{Run ForwardMole 1D Python Model from R}
\usage{
ForwardMole1D(
  workspace = ".",
  name,
  python_path = "C:/Users/marco/anaconda3/python.exe",
  python_script_path = "Data/forward_mole_1D_2.py"
)
}
\arguments{
\item{workspace}{Character. Base working directory path where output folders reside. Default is `"."`.}

\item{name}{Character. Base name prefix used to build DEM and hydrograph file names.}

\item{python_path}{Character. Path to the Python executable to use with `reticulate`. Default is `"C:/Users/marco/anaconda3/python.exe"`.}

\item{python_script_path}{Character. Path to the Python script to source. Default is `"Data/forward_mole_1D_2.py"`.}
}
\value{
Invisible `NULL`. Side effects include executing the Python FM1D function, which processes and saves outputs in the specified workspace.
}
\description{
Configures Python environment using `reticulate`, sources the ForwardMole 1D Python script,
and runs the `FM1D` function to perform 1D hydrological processing on the DEM and hydrograph data.
}
\details{
This function is a wrapper to run a Python-based hydrological model from within R. It sets the Python environment, sources the Python script,
and calls the `FM1D` function with constructed paths based on the input parameters.

Make sure the Python script `forward_mole_1D_2.py` is accessible at the specified path and the Python environment has all necessary dependencies installed.
}
\examples{
\dontrun{
ForwardMole1D(
  workspace = "C:/Projects/HydroModel",
  name = "basin01",
  python_path = "C:/Users/marco/anaconda3/python.exe",
  python_script_path = "Data/forward_mole_1D_2.py"
)
}

}
\references{
Rápalo, L. M. C., Gomes Jr, M. N., & Mendiondo, E. M. (2024).
Developing an open-source flood forecasting system adapted to data-scarce regions:
A digital twin coupled with hydrologic-hydrodynamic simulations. *Journal of Hydrology*.
https://doi.org/10.1016/j.jhydrol.2024.131929

Source code: \url{https://github.com/luiscastillo1993/Forward-Mole}
}
